pyproject.toml
[project]
name = 'scylla-cluster-tests'
version = '0.1'
requires-python = ">=3.10, <3.14"
dependencies = [
    "aexpect==1.6.2",
    "anyconfig==0.14.0",
    "awscli==1.40.34",
    "azure-identity==1.16.1",
    "azure-keyvault-keys==4.11.0",
    "azure-mgmt-compute==37.0.1",
    "azure-mgmt-keyvault==12.0.0",
    "azure-mgmt-msi==7.0.0",
    "azure-mgmt-network==30.0.0",
    "azure-mgmt-resource==24.0.0",
    "azure-mgmt-resourcegraph==8.0.0",
    "azure-mgmt-subscription==3.1.1",
    "azure-storage-blob==12.16.0",
    "boto3==1.38.35",
    "boto3-stubs[s3,ec2,dynamodb,pricing]==1.28.4",
    "click==8.1.3",
    "confluent-kafka[avro]==2.10.1",
    "deepdiff==8.6.1",
    "docker==7.1.0",
    "elasticsearch==9.2.0",
    "fabric==3.2.2",
    "fastavro==1.11.1",
    "gimme-aws-creds==2.8.2",
    "google-api-python-client==2.93.0",
    "google-auth>=2.0.0",
    "google-cloud-compute==1.13.0",
    "google-cloud-kms>=3.5.1",
    "google-cloud-storage==3.5.0",
    "hdrhistogram==0.9.2",
    "invoke==2.2.0",
    "jinja2==3.1.6",
    "jira~=3.10.0",
    "kubernetes==34.1.0",
    "ldap3-bleeding-edge==2.10.1.1338",
    "lz4>3.0.0",
    "moto[server]==5.1.1",
    "mysql-connector-python==9.1.0",
    "packaging>=24",
    "parameterized==0.8.1",
    "paramiko==4.0.0",
    "pre-commit==4.0.1",
    "prettytable==3.10.2",
    "prometheus_client==0.11.0",
    "pycryptodome==3.19.1",
    "pydantic==2.10.5",
    "PyGithub==2.1.1",
    "pytest==9.0.1",
    "pytest-random-order==1.0.4",
    "pytest-xdist==3.8.0",
    "python-jenkins==1.8.2",
    "PyYAML==6.0.1",
    "pyzmq==27.1.0",
    "questionary==2.1.0",
    "requests==2.32.4",
    "scylla-driver==3.29.5",
    "sortedcontainers==2.4.0",
    "ssh2-python>=1.1.2.post1",
    "tcconfig==0.29.1",
    "tenacity==9.0.0",
]
[dependency-groups]
dev = [
    "autopep8==2.3.2",
    "pre-commit==4.0.1",
    "pycodestyle==2.13.0",
    "ruff==0.14.6",
    "uv-sort>=0.4.0",
]

[tool.uv]
package = false
# gimme-aws-creds does not work on Windows for Python 3.9+
environments = ["sys_platform != 'win32'"]

[tool.ruff]
lint.select = [
    "BLE",
    "F401", "F821", "F823", "F841",
    "PL", "PLR0913","PLR0914", "PLR0916",
    "YTT",
    "F541",
    "PIE",
    "B006",
    "PLW", # mainly for PLW0211
]
lint.ignore = ["E501", "PLR2004"]

lint.preview = true
lint.explicit-preview-rules = true
exclude = ["argus/"]

target-version = "py313"

force-exclude = true
line-length = 240
respect-gitignore = true

[tool.ruff.lint.pylint]
max-args = 12
max-statements = 100
max-branches = 24
max-locals = 15

[tool.autopep8]
max_line_length = 120
ignore = "E226,E24,W50,W690,E402,E731"
